// src/app/model/dados-colaborador.model.ts
export class DadosColaborador {
    id_colaborador: any = '';
    nome: any = '';
    cpf: any = '';
    sexo: any = '';
    email: any = '';
    lattes: any = '';
    matricula: any = '';
    data_ingresso: any = '';
    data_nascimento: any = '';
    data_desligamento: any = '';
    dt_prev_conclusao: any = '';
    observacoes: any = '';
    ativo: any = '';
    dt_cad: any = '';
    dt_alt: any = '';
    id_user: any = '';
    id_nacionalidade: any = '';
    id_cidade_nascimento: {
        nome_cidade: any;
    };
    id_tp_vinculo: any = '';
    id_maior_formacao: any = '';
    id_user_cad: any = '';
    id_user_alt: any = '';
}